Updated final.md To Match Google Doc
Mostly matches Google Doc at this point, though is missing the formulas in there. Unsure if they are needed though with the amount that is already written. Currently testing if video and return image embedding is working. Also added FinalReturns folder to make it easier to access images for embedding.